teams_bot/requirements.txt
# Teams Bot Requirements
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
aiohttp>=3.8.0
asyncio>=3.4.3

# For authentication (if needed)
msal>=1.24.0

# Logging and utilities
python-dotenv>=1.0.0
